+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/ui_components/pin_code_keyboard.dart
@@ -72,7 +72,8 @@ class PinCodeKeyboard extends StatelessWidget {
   /// for the package
   final BiometricsLocalDataSource? biometricsLocalDataSource;
 
-  /// Callback called when the user authentication succeeds
+  /// Callback called when the user authentication succeeds. It accepts a dynamic
+  /// value which is forwarded from the `verifyPinCode` method of the [pinCodeService].
   final Function(dynamic)? onAuthenticated;
 
   /// Provide custom implementation for the most down left button, shown when